Add files via upload
The primary objective of my program is to immerse the user in an engaging game of Score Four! To emerge victorious in a round, players must secure two game wins by arranging four of their icons consecutively. The game board comprises 16 squares, each accompanied by a textual attribute, representing the available options for users to select during gameplay.

Upon pressing a button, the program proceeds through the onEvent function, where it checks for any prior moves in the chosen square. If a move has already been made in that square, the user is prompted to select a different square. However, if the square is unoccupied, the program proceeds to place the respective player's text (designated as 'P1' for Player 1 and 'P2' for Player 2) in the square, signaling the start of the next player's turn.

Whenever a player clicks on a square, their designated player text (either 'P1' or 'P2') is deposited into the square. Additionally, users have the option to reset the game board after each match, effectively removing all previously placed text. Furthermore, if players wish to start over entirely, they can reset the score as well.